backend/requirements.txt
# Core server
fastapi==0.128.0
uvicorn==0.40.0
gunicorn==23.0.0
python-multipart==0.0.6

# OpenAI
openai==1.54.0

# Vector DB
chromadb==0.4.22

# Embeddings & LLM orchestration
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Document ingestion
pypdf==3.17.4
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
beautifulsoup4==4.12.2
requests==2.31.0

# Env & CORS
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.1.0

# Additional dependencies
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
h11==0.16.0
idna==3.11
starlette==0.50.0
typing-extensions==4.15.0
httpx==0.27.0